Change report to use size_t instead of int

+++ interpreter/include/interpreter.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <token.h>
+
 #include <string>
 
 void runFile(const std::string&);
@@ -8,5 +10,6 @@ void runPrompt();
 
 void run(const std::string&);
 
-void error(int, const std::string&);
-void report(int, const std::string&, const std::string&);
+void error(size_t, const std::string&);
+
+void report(size_t, const std::string&, const std::string&);
